Refactoring test approach
