+++ test/contracts/SelfDestructible.js
