+++ test/contracts/index.js
